requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
api==0.0.7
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
bcrypt==4.0.1
black==22.12.0
bson==0.5.10
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
dnspython==2.3.0
docutils==0.20.1
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.96.0
filelock==3.12.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.0.3
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.3
motor==3.1.2
multidict==6.0.4
mypy-extensions==1.0.0
newpy==1.1.16
nose==1.3.7
packaging==21.3
parsel==1.8.1
passlib==1.7.4
pathspec==0.10.3
Pillow==9.4.0
pip-autoremove==0.10.0
pluggy==1.2.0
platformdirs==2.6.2
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
PyDispatcher==2.0.7
PyJWT==2.8.0
pymongo==4.3.3
pyOpenSSL==23.2.0
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
queuelib==1.6.2
requests-file==1.5.1
service-identity==21.1.0
six==1.15.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tldextract==3.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
w3lib==2.1.1
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zope.interface==6.0
